requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3

# Data Visualization
plotly==5.18.0
altair==5.2.0

# Database
pymongo==4.6.1
sqlalchemy==2.0.25

# Reddit API
praw==7.7.1

# Machine Learning & NLP
transformers==4.37.2
torch>=2.2.0
scikit-learn==1.4.0
scipy==1.12.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
python-dateutil==2.8.2

# Data Processing
beautifulsoup4==4.12.3
lxml==5.1.0

# Logging
loguru==0.7.2

#Youtube API
google-api-python-client==2.187.0


# Optional: Topic Modeling
gensim==4.3.2